feat(mvf-v3): adversarially-reviewed score redesign + IF-SI experiment tooling
Score v3 (docs/plan_low_quality_multiview.md v3, after a 12-agent
adversarial review found the v2 gate quantitatively broken):
- Q: zero-anchored calibrated sigmoid gate clip(2(σ(ΔL/s)−0.5),0,1),
  s = P10(ΔL_clean)/logit(0.8) from a clean reference
  (scripts/calibrate_reliability.py); rank01 kept as v1 ablation;
  K>1 counterfactuals -> mean-of-gates with dispersion discount
- D: progress ranked within Selected(t-1) only (rich-get-richer fix),
  unselected neutral; d_floor=0.5 range compression so the
  learnability factor (~2x) can no longer override the gate (~61x)
- A: pool-CDF (rank01) normalisation in MVF mode + collapse guard;
  adaptive λ_t = λ0 · eigengap-weighted anchor stability
- rank01: midrank ties (dataset order no longer leaks into scores)
- completeness: raw-text heuristic via text_complete column — the
  EOS-only check provably could not detect T3 truncation
- hard error on missing reliability cache at epoch>1 (both in
  _prepare_mvf before the counterfactual forward and in selector);
  cache re-gates from stored losses on config change
- fp32 CE cast in compute_rewards (ΔL was mixing bf16/fp32 exactly
  at the zero-anchor kink)

Tooling:
- scripts/make_table_v2.py: one row = one sealed run (max-of-each is
  unrepresentable), seed-macro stats, paired t-CI + exact sign-flip
  permutation + Holm, seed=None and duplicate-run_dir integrity aborts
- corruption: T1b cross-source mismatch, T7 fluent-wrong (two-step,
  PENDING pool naming), K counterfactuals; byte-identical legacy
  output proven by SHA256 regression tests
- scripts/{compute_uncond_loss,make_oracle_pool,gen_fluent_wrong,
  calibrate_reliability}.py; score_pool gains Q_rank/mvf_v2/additive/
  ppl/ifd rows (IFD>1 exclusion per Li et al.)
- lowq arm configs share one training fragment (method-first merge
  order; resolution pinned by tests/test_lowq_configs.py)
- config loader: utf-8; env empty-string scale no longer crashes MVF

Docs: plan v3 (positioning, γ* margins 61x/3.96x, pre-registered
stats protocol), CIKM audit archived, runbook banner.

Tests: 125 passed (was 48).

+++ tads/core/reward.py
@@ -67,8 +67,13 @@ def compute_rewards(
         rm = resp_mask[i]                                     # (T-1,)
         nr = n_resp[i]
 
-        # CE over response tokens only.
-        ce_i = F.cross_entropy(sl, ll, reduction="none")      # (T-1,) fp32
+        # CE over response tokens only. Cast to fp32 BEFORE the CE — the
+        # comment previously claimed fp32 but no cast happened, so
+        # loss_orig carried bf16 rounding (~0.4% rel.) while the
+        # counterfactual pass (reliability.compute_pool_loss) is fp32.
+        # ΔL = loss_cf − loss_orig is zero-ANCHORED at the rezero kink,
+        # exactly where that dtype asymmetry moves samples across Q = 0.
+        ce_i = F.cross_entropy(sl.float(), ll, reduction="none")  # (T-1,) fp32
         r_loss[i] = (ce_i * rm).sum() / nr
 
         # Entropy via log_softmax: H = -Σ p log p = -Σ exp(lp) * lp.
